requirements.txt
altair
beautifulsoup4
black
click
interpret
jupyter
jupyterlab
kaggle
lightgbm
line-profiler
missingno
mlflow
numpy
orderedmultidict
pandas
pandas-profiling
path.py
plotly
pyarrow
scikit-learn
scipy
seaborn
snakeviz
shap
statsmodels
umap-learn
vega
xgboost